확장을 위해서 Service 클래스 생성 및 빈 설정
