+++ todo/src/main/java/com/todo/todo/Controller/TodoController.java
@@ -2,6 +2,7 @@ package com.todo.todo.Controller;
 
 import com.todo.todo.Entity.Todo;
 import com.todo.todo.Repository.TodoRepository;
+import com.todo.todo.Service.TodoService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,23 +11,24 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/todos")
 public class TodoController {
-    private final TodoRepository todoRepository;
 
-    public TodoController(TodoRepository todoRepository) {
-        this.todoRepository = todoRepository;
+    private final TodoService todoService;
+
+    public TodoController(TodoService todoService) {
+        this.todoService = todoService;
     }
 
     // ✅ 할 일 추가
     @PostMapping
     public ResponseEntity<Todo> createTodo(@RequestBody Todo todo) {
-        Todo savedTodo = todoRepository.save(todo);
+        Todo savedTodo = todoService.createTodo(todo);
         return ResponseEntity.ok(savedTodo);
     }
 
     // ✅ 할 일 목록 조회
     @GetMapping
     public ResponseEntity<List<Todo>> getTodos() {
-        List<Todo> todos = todoRepository.findAll();
+        List<Todo> todos = todoService.findAll();
         return ResponseEntity.ok(todos);
     }
 }
